wrap hash calculation in curly brackets.
Hello everyone,

I'm compiling the chain branch, and I'm getting warnings all the time for this line. I think the author meant to run the two statements in the for loop body and not just one, since otherwise the factor variable would be unnecessary. 

If this is not the case please ignore this pull request. Thanks.